Now the game is playable with two people, Missing the AI for the enemy Fixing a small bug with the point system (the game is over when one reaches 5 (but the score shows only 4, but it is 5 "internal" Points.

+++ src/Pong/GameObject.java
@@ -7,7 +7,7 @@ public abstract class GameObject {
 
     protected int x, y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
     public GameObject(int x, int y, ID id){
         this.x = x;
@@ -47,7 +47,7 @@ public abstract class GameObject {
         this.velX = velX;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
@@ -55,7 +55,7 @@ public abstract class GameObject {
         this.velY = velY;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
